Convert type system to u32/i32 for RV32 address space
The assembler targets a 32-bit architecture where addresses and sizes fit
naturally in 32-bit values. Convert i64/u64 types to i32/u32 throughout to
improve type correctness and clarity.

Key changes:
- encode_source_with_size_tracking(): return type i64 -> u32 for bss_size
- converge_and_encode(): return type cascades from encoder.rs
- dump.rs: format_address() and get_segment_base() now use u32
- main.rs: remove unnecessary casts, simplify arithmetic

All address calculations now use u32 natively, casting only when needed
(e.g., for display). Expression checking functions retain i64 for
overflow detection during immediate validation.

All 227+ tests pass. No functional changes - purely type system improvements.

+++ src/assembler.rs
@@ -155,7 +155,7 @@ pub fn converge_and_encode<C: ConvergenceCallback>(
     text_start: u32,
     callback: &C,
     show_progress: bool,
-) -> Result<(Vec<u8>, Vec<u8>, i64)> {
+) -> Result<(Vec<u8>, Vec<u8>, u32)> {
     const MAX_ITERATIONS: usize = 10;
 
     if show_progress {
